[HMTM-187] - Refactoring/tests update
