+++ src/common/dependencies.py
@@ -1,16 +1,18 @@
 from typing import Optional, List, Dict
 
-from starlette.requests import Request
-
-from graphql_client import GetMeQuery, extract_error_message
+from fastapi.requests import Request
+from graphql_client import (
+    GetMeQuery,
+    extract_error_message,
+    RefreshTokensMutation,
+    ResponseProcessor as GQLResponseProcessor
+)
 from graphql_client.dto import GQLResponse
-from src.config import config
-from src.constants import DEFAULT_ERROR_MESSAGE
-from src.cookies import CookiesConfig
-from src.sso.constants import ERRORS_MAPPING
-from src.sso.dependencies import refresh_tokens
-from src.sso.dto import GetMeResponse, RefreshTokensResponse
-from src.sso.utils import user_from_dict
+from src.common.config import config
+from src.common.constants import DEFAULT_ERROR_MESSAGE, ERRORS_MAPPING
+from src.common.cookies import CookiesConfig
+from src.common.dto import GetMeResponse, RefreshTokensResponse
+from src.common.utils import user_from_dict
 
 
 async def get_me(
@@ -45,8 +47,7 @@ async def get_me(
             refreshed_tokens: RefreshTokensResponse = await refresh_tokens(request=request, cookies=cookies)
 
             if refreshed_tokens.error is not None:
-                result.error = refreshed_tokens.error
-                return result
+                raise Exception(refreshed_tokens.error)
 
             result.headers = refreshed_tokens.headers
             result.cookies = refreshed_tokens.cookies
@@ -62,8 +63,7 @@ async def get_me(
             )
 
             if "errors" in gql_get_me.result:
-                result.error = gql_get_me.result["errors"][0]["message"]
-                return result
+                raise Exception(gql_get_me.result["errors"][0]["message"])
 
             result.user = user_from_dict(gql_get_me.result["me"])
 
@@ -76,5 +76,46 @@ async def get_me(
                 DEFAULT_ERROR_MESSAGE
             )
             result.error = error
+
+    return result
+
+
+async def refresh_tokens(
+        request: Request,
+        cookies: Optional[List[CookiesConfig]] = None
+) -> RefreshTokensResponse:
+    result: RefreshTokensResponse = RefreshTokensResponse()
+
+    actual_cookies: Dict[str, str] = request.cookies
+    if cookies:
+        for cookie in cookies:
+            actual_cookies[cookie.KEY] = cookie.VALUE
+
+    try:
+        gql_refresh_tokens: GQLResponse = await config.graphql_client.gql_query(
+            query=RefreshTokensMutation.to_gql(),
+            variable_values={},
+            cookies=actual_cookies
+        )
+
+        if "errors" in gql_refresh_tokens.result:
+            raise Exception(gql_refresh_tokens.result["errors"][0])
+
+        result.result = True
+        result.headers = gql_refresh_tokens.headers  # type: ignore[assignment]
+
+        if gql_refresh_tokens.headers is not None:
+            result.cookies = GQLResponseProcessor(gql_response=gql_refresh_tokens).get_cookies()
+
+    except Exception as err:
+        error: str = ERRORS_MAPPING.get(
+            extract_error_message(
+                error=str(err),
+                default_message="Ошибка обновления токенов"
+            ),
+            DEFAULT_ERROR_MESSAGE
+        )
+
+        result.error = error
 
     return result
