--- src/constants.py
@@ -1,1 +1,0 @@
-DEFAULT_ERROR_MESSAGE: str = "Неизвестная ошибка"
