+++ src/root/routes.py
@@ -4,10 +4,10 @@ from fastapi import APIRouter, Request, Depends
 from fastapi.responses import HTMLResponse, Response
 from fastapi.templating import Jinja2Templates
 
-from src.cookies import set_cookie
-from src.sso.dto import GetMeResponse
+from src.common.cookies import set_cookie
+from src.common.dto import GetMeResponse
 from src.common.dependencies import get_me as get_me_dependency
-from src.utils import extract_url_error_message
+from src.common.utils import extract_url_error_message
 
 main_router = APIRouter(tags=["Main"])
 templates = Jinja2Templates(directory="templates")
