+++ src/sso/dto.py
@@ -2,13 +2,8 @@ from dataclasses import dataclass
 from typing import Optional, List
 
 from src.sso.models import UserInfo
-from src.profile.models import User, Master
-from src.dto import BaseResponse, BoolResponse
-
-
-@dataclass
-class GetMeResponse(BaseResponse):
-    user: Optional[User] = None
+from src.profile.models import Master
+from src.common.dto import BoolResponse
 
 
 @dataclass
@@ -18,6 +13,11 @@ class RegisterResponse(BoolResponse):
 
 @dataclass
 class LoginResponse(BoolResponse):
+    pass
+
+
+@dataclass
+class ChangeForgetPasswordResponse(BoolResponse):
     pass
 
 
@@ -33,11 +33,6 @@ class SendVerifyEmailMessageResponse(BoolResponse):
 
 @dataclass
 class SendForgetPasswordMessageResponse(BoolResponse):
-    pass
-
-
-@dataclass
-class RefreshTokensResponse(BoolResponse):
     pass
