+++ src/toys/routes.py
@@ -2,9 +2,9 @@ from fastapi.responses import HTMLResponse, RedirectResponse
 from fastapi import Response, Request, Depends, status, APIRouter
 from starlette.templating import Jinja2Templates
 
-from src.cookies import set_cookie
-from src.utils import FernetEnvironmentsKey, encryptor as encryptor_dependency
-from src.sso.dto import GetMeResponse
+from src.common.cookies import set_cookie
+from src.common.utils import FernetEnvironmentsKey, encryptor as encryptor_dependency
+from src.common.dto import GetMeResponse
 from src.toys.dto import ToysCatalogResponse, ToyByIDResponse
 from src.common.dependencies import get_me as get_me_dependency
 from src.toys.dependencies import (
